Add customMetrics to network manager
Pass the browser custom metrics to the network manager. The network
manager will use these metrics instead of the built in ones. This will
enable us to differentiate between k6 and browser metrics.

We could just retrieve the customMetrics from the ctx too, but this
feels a bit more explicit which is easier to understand.