Initialize browser on newPage
The browser.newPage() method will automatically initialize or retrieve a
previously initialized browser for the scenario and  iteration. This
browser is held at the root module level pool throughout the iteration
lifecycle. When the iteration ends, signaled by the VU context being
done, the browser is closed and removed from the pool.