Add k6ext.GetScenarioName
Retrieves the scenario name for the given VU context.
This is required in order to build the key for the browsers pool, as the
VU iteration is just a monotonic integer that is shared per scenario.
Therefore, if we don't want to share the browser instance per scenario,
we have to build the pooll key based on the scenario name also.